Refactors JobFileLink into its own function
